refactor(email_service/send_email): split download_file_content function into file_util.py (SCRUM-73)
