+++ src/email_service/send_email/file_util.py
@@ -1,0 +1,13 @@
+import requests
+
+
+def download_file_content(file_url):
+    """
+    Download the content of the file from the given URL.
+
+    :param file_url: URL of the file to be downloaded.
+    :return: Content of the downloaded file.
+    """
+    response = requests.get(file_url, timeout=25)
+    response.raise_for_status()
+    return response.content
